src/backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.1
django-cors-headers==4.3.1
django-email==0.1.10
django-otp==1.3.0
django-sslserver==0.22
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
style==1.1.0
typing_extensions==4.9.0
update==0.0.1
uritemplate==4.1.1
urllib3==2.2.0
